+++ ts2dart/lib/src/ast/types/static.dart
@@ -100,7 +100,6 @@ enum InteropStaticType with InteropType, WithInteropTypeParams {
       symbol: 'Float64List',
       mappings: {'Float64Array'},
       package: 'dart:typed_data'),
-  futureOr(symbol: 'FutureOr', package: 'dart:async', typeParamsLength: 1),
   banned(symbol: 'Never', mappings: {'intrinsic'}),
   never(symbol: 'Never', mappings: {'never'}, typeParamsLength: 0),
   ctor(symbol: 'Never', mappings: {'constructor'}, typeParamsLength: 0);
@@ -250,9 +249,6 @@ enum InteropStaticType with InteropType, WithInteropTypeParams {
 
   @override
   bool get optional => isDynamic;
-
-  @override
-  bool get isPromiseLike => this == futureOr;
 
   @override
   Reference ref(